Added localStorage aware logged in user
Current User

It would be bothersome if the user had to log in every time they open
your app. You can avoid this by using the cached current Parse.User
object.

Whenever you use any signup or login methods, the user is cached in
localStorage. You can treat this cache as a session, and automatically
assume the user is logged in: